minor adjustment regarding quotes
Made it so quotes that wrap the entire input are removed from word length totals. This may cause issues so I want to add a toggle in the future. This was added to make it easier to copy and paste data from a spreadsheet without having to tweak it each time.